pyproject.toml
[build-system]
requires = ['setuptools']
build-backend = 'setuptools.build_meta'

[project]
name = 'pyvista'
description = 'Easier Pythonic interface to VTK'
authors = [
    {name = 'PyVista Developers', email = 'info@pyvista.org'},
]
readme = 'README.rst'
requires-python = '>=3.8'
keywords = ['vtk', 'numpy', 'plotting', 'mesh']
license = {text = 'MIT'}
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Intended Audience :: Science/Research',
    'Topic :: Scientific/Engineering :: Information Analysis',
    'License :: OSI Approved :: MIT License',
    'Operating System :: Microsoft :: Windows',
    'Operating System :: POSIX',
    'Operating System :: MacOS',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]
dependencies = [
    'imageio',
    'matplotlib>=3.0.1',
    'numpy',
    'pillow',
    'pooch',
    'scooby>=0.5.1',
    'vtk',  # keep without version constraints
]
dynamic = ['version']

[project.optional-dependencies]
all = [
    'cmocean',
    'colorcet',
    'ipyvtklink',
    'meshio>=5.2',
    'panel',
    'pythreejs',
]
colormaps = [
    'cmocean',
    'colorcet',
]
io = ['meshio>=5.2']
jupyter = [
    'ipyvtklink',
    'ipywidgets',
    'jupyter-server-proxy',
    'nest_asyncio',
    'panel',
    'pythreejs',
    'trame>=2.2.6',
    'trame-client>=2.4.2',
    'trame-server>=2.8.0',
    'trame-vtk>=2.4.0',
]
trame = [
    'trame>=2.2.6',
    'trame-client>=2.4.2',
    'trame-server>=2.8.0',
    'trame-vtk>=2.4.0',
]

[project.urls]
Documentation = 'https://docs.pyvista.org/'
"Bug Tracker" = 'https://github.com/pyvista/pyvista/issues'
"Source Code" = 'https://github.com/pyvista/pyvista'

[tool.setuptools.dynamic]
version = {attr = 'pyvista._version.__version__'}

[tool.setuptools.packages.find]
include = [
    'pyvista',
    'pyvista.*',
]

[tool.setuptools.package-data]
pyvista = [
    'py.typed',
]
"pyvista.examples" = [
    '2k_earth_daymap.jpg',
    'airplane.ply',
    'ant.ply',
    'channels.vti',
    'globe.vtk',
    'hexbeam.vtk',
    'nut.ply',
    'rectilinear.vtk',
    'sphere.ply',
    'uniform.vtk',
]

[tool.ruff.isort]
# Sort by name, don't cluster "from" vs "import"
force-sort-within-sections = true
# Combines "as" imports on the same line
combine-as-imports = true

[tool.black]
line-length = 100
skip-string-normalization = true
target-version = ['py39']
exclude='\.eggs|\.git|\.mypy_cache|\.tox|\.venv|_build|buck-out|build|dist|node_modules'

[tool.blackdoc]
# From https://numpydoc.readthedocs.io/en/latest/format.html
# Extended discussion: https://github.com/pyvista/pyvista/pull/4129
# The length of docstring lines should be kept to 75 characters to facilitate
# reading the docstrings in text terminals.
line-length = 75

[tool.build_sphinx]
source-dir = 'doc'
build-dir  = './doc/_build'
all_files  = 1

[tool.upload_sphinx]
upload-dir = 'doc/_build/html'

[tool.pydocstyle]
match = '(?!coverage).*.py'
convention = "numpy"
add-ignore = ["D404"]

[tool.codespell]
skip = '*.pyc,*.txt,*.gif,*.png,*.jpg,*.ply,*.vtk,*.vti,*.vtu,*.js,*.html,*.doctree,*.ttf,*.woff,*.woff2,*.eot,*.mp4,*.inv,*.pickle,*.ipynb,flycheck*,./.git/*,./.hypothesis/*,*.yml,doc/_build/*,./doc/images/*,./dist/*,*~,.hypothesis*,./doc/examples/*,*.mypy_cache/*,*cover,./tests/tinypages/_build/*,*/_autosummary/*'
ignore-words = "doc/styles/Vocab/pyvista/accept.txt"
quiet-level = 3

[tool.coverage.run]
omit = [
    'pyvista/ext/coverage.py',
    'pyvista/conftest.py',
    # kept for backwards compatibility:
    'pyvista/plotting/theme.py',
]

[tool.pytest.ini_options]
junit_family='legacy'
filterwarnings = [
    'ignore::FutureWarning',
    'ignore::PendingDeprecationWarning',
    'ignore::DeprecationWarning',
    # bogus numpy ABI warning (see numpy/#432)
    'ignore:.*numpy.dtype size changed.*:RuntimeWarning',
    'ignore:.*numpy.ufunc size changed.*:RuntimeWarning',
    'ignore:.*Given trait value dtype "float64":UserWarning',
    'ignore:.*The NumPy module was reloaded*:UserWarning',
    'error::pyvista.PyVistaDeprecationWarning'
]
doctest_optionflags = 'NUMBER ELLIPSIS'
testpaths = 'tests'
markers = [
    'needs_vtk_version(version): skip test unless VTK version is at least as specified.',
    'needs_download: this test downloads data during execution',
]
image_cache_dir = "tests/plotting/image_cache"

[tool.mypy]
ignore_missing_imports = true

[tool.ruff]
target-version = "py38"
select = [
    # isort
    "I",
]
fixable = [
    # isort
    "I",
]

[tool.ruff.per-file-ignores]
'pyvista/__init__.py' = ["I001"]
'pyvista/**/__init__.py' = ["I001"]
